Destructurize game engine
