+++ packages/client/src/types/gif.d.ts
@@ -1,0 +1,1 @@
+declare module 'gif'
